test: split tests by source module
